+++ app/assets/javascripts/components/story/StoryStateButtons.js
@@ -5,18 +5,13 @@ const renderInputButton = (event) => (
     type="button"
     key={event}
     className={`transition ${event}`}
-    value={I18n.t('story.events.' + event)}
+    value={I18n.t(`story.events.${event}`)}
   />
 );
 
-const StateButtons = ({ events, isSearchResult }) => (
+const StateButtons = ({ events }) => (
   <div className="state-actions">
     { events.map(renderInputButton) }
-    { isSearchResult &&
-      <button id="locate" type="button" className="btn btn-default locate-btn locate">
-        <i className="mi md-gps-fixed md-14">gps_fixed</i>
-      </button>
-    }
   </div>
 );