+++ app/assets/javascripts/views/story_view.js
@@ -12,6 +12,7 @@ import StoryTasks from 'components/story/StoryTasks';
 import TaskForm from 'components/tasks/TaskForm';
 import StoryAttachment from 'components/story/StoryAttachment';
 import StoryStateButtons from 'components/story/StoryStateButtons';
+import StoryEstimateButtons from 'components/story/StoryEstimateButtons';
 
 var Clipboard = require('clipboard');
 
@@ -85,7 +86,6 @@ module.exports = FormView.extend({
     "click .cancel": "cancelEdit",
     "click .clone-story": "cloneStory",
     "click .transition": "transition",
-    "click .state-actions .estimate": "estimate",
     "change select.story_type": "render",
     "click .destroy": "clear",
     "click .description": "editDescription",
@@ -192,10 +192,10 @@ module.exports = FormView.extend({
     });
   },
 
-  estimate: function(ev) {
+  estimate: function(points) {
     this.saveInProgress = true;
     this.render();
-    this.model.set({estimate: ev.target.attributes['data-value'].value});
+    this.model.set({ estimate: points });
 
     var that = this;
     this.model.save(null, {
@@ -501,9 +501,19 @@ module.exports = FormView.extend({
         ReactDOM.render(
           <StoryStateButtons
             events={this.model.events()}
-            isSearchResult={this.isSearchResult}
           />,
           stateButtons
+        );
+      }
+
+      const estimateButtons = this.$('[data-story-estimate-buttons]').get(0)
+      if(estimateButtons) {
+        ReactDOM.render(
+          <StoryEstimateButtons
+            points={this.model.point_values()}
+            onClick={this.estimate}
+          />,
+          estimateButtons
         );
       }